refactor: organize services on specific folders
