+++ src/services/index.ts
@@ -1,4 +1,4 @@
-export { Mine as MineService } from "./mine";
-export { Nodes as NodesService } from "./nodes";
-export { Registry as RegistryService } from "./registry";
-export { Transactions as TransactionService } from "./transactions";
+export { Mine as MineService } from "./mine/mine";
+export { Nodes as NodesService } from "./nodes/nodes";
+export { Registry as RegistryService } from "./registry/registry";
+export { Transactions as TransactionService } from "./transactions/transactions";
