+++ src/services/registry/registry.unit.test.ts
